% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_id_to_job_ids.R
\name{batch_id_to_job_ids}
\alias{batch_id_to_job_ids}
\title{Get job IDs from a batch ID}
\usage{
batch_id_to_job_ids(batch_id)
}
\arguments{
\item{batch_id}{character or numeric, the ID assigned to a batch of jobs by
[run_comrad_sim_hpc()].}
}
\description{
Reads the logbook and returns the job IDs associated with the input batch ID.
}
\author{
Théo Pannetier
}
