% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_ml_log_hpc.R
\name{cat_ml_log_hpc}
\alias{cat_ml_log_hpc}
\title{Fetch and print dd ML log from Peregrine}
\usage{
cat_ml_log_hpc(job_id)
}
\arguments{
\item{job_id}{eight-digit job ID given by Peregrine upon
submission.}
}
\description{
Given a job ID, prints the corresponding `.log` file from Peregrine.
}
\author{
Théo Pannetier
}
