% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_sim_hpc.R, R/is_sim_on_hpc.R
\name{cat_sim_hpc}
\alias{cat_sim_hpc}
\alias{cat_sim_log_hpc}
\alias{cat_sim_csv_hpc}
\alias{is_sim_csv_on_hpc}
\alias{is_sim_log_on_hpc}
\title{Fetch and print simulation log from Peregrine}
\usage{
cat_sim_log_hpc(job_id, pkg = "comrad")

cat_sim_csv_hpc(job_id, pkg = "comrad")

is_sim_csv_on_hpc(job_id, pkg = "comrad")

is_sim_log_on_hpc(job_id, pkg = "comrad")
}
\arguments{
\item{job_id}{eight-digit job ID given by Peregrine upon
submission.}
}
\description{
Given a job ID, prints the corresponding `.log` file from Peregrine.
}
\author{
Théo Pannetier
}
