% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_sim_output_hpc.R
\name{cat_sim_output_hpc}
\alias{cat_sim_output_hpc}
\title{Fetch and print simulation output file from Peregrine}
\usage{
cat_sim_output_hpc(job_id)
}
\arguments{
\item{job_id}{eight-digit job ID given by Peregrine upon
submission.}
}
\description{
Given a job ID, prints the corresponding `.csv` file from Peregrine.
}
\author{
Théo Pannetier
}
