% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_job_ids.R
\name{catch_job_ids}
\alias{catch_job_ids}
\title{Submit jobs and catch their IDs}
\usage{
catch_job_ids(fun, ...)
}
\arguments{
\item{fun}{a function submitting jobs to Peregrine. The function should
otherwise be absolutely silent, i.e. the only output sent to console should
be `Submitted batch job XXXXXXXX` from Peregrine.}

\item{...}{arguments to be passed to `fun`}
}
\description{
Calls `fun`, catches job submission messages and extracts job IDs from them.
}
\author{
Théo Pannetier
}
