% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_comrad_versions.R
\name{compare_comrad_versions}
\alias{compare_comrad_versions}
\title{Compare installed comrad versions between HPC and local computer}
\usage{
compare_comrad_versions()
}
\description{
Calls [utils::packageVersion()] on the two instances and returns an error
if the two are different
}
\author{
Théo Pannetier
}
