% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_comsie_versions.R
\name{compare_comsie_versions}
\alias{compare_comsie_versions}
\title{Compare installed comsie versions between HPC and local computer}
\usage{
compare_comsie_versions()
}
\description{
Calls [utils::packageVersion()] on the two instances and returns an error
if the two are different
}
\author{
Théo Pannetier
}
