% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_logbook_entries.R
\name{complete_logbook_entries}
\alias{complete_logbook_entries}
\title{Complete logbook entries after job completion}
\usage{
complete_logbook_entries(job_ids)
}
\arguments{
\item{job_ids}{eight-digit job IDs given by Peregrine upon
submission.}
}
\description{
`status` and `runtime` entries can be written only after a job has completed
(or failed). This function download the logbook from Peregrine, writes the
entries for `job_ids`, and uploads back to Peregrine.
}
\author{
Théo Pannetier
}
