% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comrad_params_grid.R
\name{comrad_params_grid}
\alias{comrad_params_grid}
\title{Shortcut for grid of sigma_alpha, sigma_k values}
\usage{
comrad_params_grid(add_is_retained = FALSE)
}
\description{
Shortcut for grid of sigma_alpha, sigma_k values
}
