% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comsie_params_grid.R
\name{comsie_params_grid}
\alias{comsie_params_grid}
\title{Shortcut for grid of sigma_alpha, gamma values}
\usage{
comsie_params_grid()
}
\description{
Shortcut for grid of sigma_alpha, gamma values
}
