% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_to_hd.R
\name{copy_sim_log_to_hd}
\alias{copy_sim_log_to_hd}
\title{Copy comrad files to hard drive}
\usage{
copy_sim_log_to_hd(job_ids, overwrite = FALSE)
}
\arguments{
\item{job_ids}{eight-digit job ID given by Peregrine upon submission.}

\item{overwrite}{logical, passed to [fs::file_copy()].}
}
\description{
Copy comrad files to hard drive
}
\author{
Theo Pannetier
}
