% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_comsie_params.R
\name{create_comsie_params}
\alias{create_comsie_params}
\title{Create a list of standard comsie parameters}
\usage{
create_comsie_params(
  immigration_rate = 0.001,
  mainland_nb_species = comsie::default_mainland_nb_species(),
  mainland_z_sd = comsie::default_mainland_z_sd(),
  competition_sd = comsie::default_competition_sd(),
  carrying_cap_sd = comsie::default_carrying_cap_sd(),
  carrying_cap_opt = comrad::default_carrying_cap_opt(),
  trait_opt = comrad::default_trait_opt(),
  growth_rate = comrad::default_growth_rate(),
  mutation_sd = comrad::default_mutation_sd(),
  trait_dist_sp = comrad::default_trait_dist_sp()
)
}
\arguments{
\item{competition_sd}{numeric `>= 0`.Parameter \eqn{\sigma_{\alpha}} of
the competition coefficient. Modulates the the strength of competition
between two individuals given their distance in trait space.}

\item{carrying_cap_sd}{numeric `>= 0`. Parameter \eqn{\sigma_K} of the
carrying capacity. Modulates how fast the carrying capacity decays when
moving away from the optimal trait value.}

\item{carrying_cap_opt}{numeric, value of the carrying capacity at `trait_opt`}

\item{trait_opt}{numeric. The optimal trait value, \eqn{z_{opt}}.
\code{get_carrying_cap(trait_opt) = carrying_cap_opt}.}

\item{growth_rate}{numeric `>= 0`, the baseline growth rate. Generations
being discrete, high values will cause chaos.}

\item{mutation_sd}{numeric `>= 0`, the standard deviation of the normal
distrbution from which mutations are drawn.}

\item{trait_dist_sp}{numeric, the minimal trait distance between two
clusters of individuals triggering speciation.}
}
\description{
Just returns a list with all standard parameters. This makes it easy to
modify a few while keeping default values for others.
}
\author{
Théo Pannetier
}
