% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dd_ml_with_fossil_hpc.R
\name{download_data_with_fossil_hpc}
\alias{download_data_with_fossil_hpc}
\alias{download_dd_ml_with_fossil_rds_hpc}
\alias{download_dd_ml_with_fossil_log_hpc}
\title{Download DD ML result files from Peregrine}
\usage{
download_dd_ml_with_fossil_rds_hpc(job_ids)

download_dd_ml_with_fossil_log_hpc(job_ids)
}
\arguments{
\item{job_ids}{8-digit numeric or character vector. Peregrine job IDs
identifying the files to download.}
}
\description{
Given a vector of job IDs, will download the corresponding `.rds` or `.log`
DD ML with fossil results files from Peregrine to the local instance of
`fabrika`.
}
\author{
Théo Pannetier
}
