% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_logbook_hpc.R
\name{download_logbook_hpc}
\alias{download_logbook_hpc}
\title{Download the logbook from Peregrine to local}
\usage{
download_logbook_hpc(which_one = "sims")
}
\arguments{
\item{which_one}{character, which logbook? can be `"sims"` or `"dd_ml_without_fossil"`}
}
\description{
Download the logbook from Peregrine to local
}
\author{
Théo Pannetier
}
