% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_sim.R
\name{download_sim}
\alias{download_sim}
\alias{download_sim_output}
\alias{download_sim_log}
\title{Download simulation output from Peregrine}
\usage{
download_sim_output(job_ids)

download_sim_log(job_ids)
}
\arguments{
\item{job_ids}{8-digit numeric or character vector. The IDs of jobs to
get the output of.}
}
\description{
Given a vector of job IDs, will download the corresponding `.csv` output or `.log`
files from Peregrine to the corresponding `comrad` directory.
}
\author{
Théo Pannetier
}
