% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dd_ml_without_fossil_hpc.R,
%   R/download_sim_hpc.R
\name{download_sim_hpc}
\alias{download_sim_hpc}
\alias{download_dd_ml_without_fossil_rds_hpc}
\alias{download_dd_ml_without_fossil_log_hpc}
\alias{download_sim_csv_hpc}
\alias{download_sim_log_hpc}
\title{Download DD ML result files from Peregrine}
\usage{
download_dd_ml_without_fossil_rds_hpc(job_ids)

download_dd_ml_without_fossil_log_hpc(job_ids)

download_sim_csv_hpc(job_ids)

download_sim_log_hpc(job_ids)
}
\arguments{
\item{job_ids}{8-digit numeric or character vector. Peregrine job IDs
identifying the files to download.}
}
\description{
Given a vector of job IDs, will download the corresponding `.csv` or `.log`
files from Peregrine to the local instance of `fabrika`.

Given a vector of job IDs, will download the corresponding `.csv` or `.log`
files from Peregrine to the local instance of `fabrika`.
}
\author{
Théo Pannetier

Théo Pannetier
}
