% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ggplot_data.R
\name{extract_ggplot_data}
\alias{extract_ggplot_data}
\title{Extract data from a ggplot}
\usage{
extract_ggplot_data(gg)
}
\arguments{
\item{gg}{the output of `ggplot2::ggplot()`}
}
\value{
a tibble
}
\description{
Just so I don't end up looking it up for 2 hrs again
}
