% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_job_ids_console.R
\name{extract_job_ids_console}
\alias{extract_job_ids_console}
\title{Extract job IDs from console output}
\usage{
extract_job_ids_console(console_output)
}
\arguments{
\item{console_output}{character string, the message printed to console
after submitting jobs, e.g. via [run_comrad_sim_hpc] or directly via
`sbatch`. Should be of the form `"(Submitted batch job [:digit:]{8}\\n)+"`}
}
\description{
Reads the output printed to console after submitting jobs on Peregrine
and extract all job IDs found in there.
}
\author{
Théo Pannetier
}
