% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_aic_best2.R
\name{filter_aic_best2}
\alias{filter_aic_best2}
\title{Filter best ML results and compute AIC-based selection scores for `dd_ML` results without fossil}
\usage{
filter_aic_best2(ml_tbl)
}
\arguments{
\item{mle_tbl}{a data frame containing maximum likelihood estimates for each of
the DD models (at least one row for each model), with at least cols `dd_model`,
`tree` and `aic`. E.g. the output of [read_mle_tbl_without_fossil()].}
}
\description{
Returns a table with best maximum likelihood results for each of the 10 DD
models, and computes the delta AIC, AIC weights, and the latter's cumulative sum.
}
