% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_job_completed.R
\name{has_job_completed}
\alias{has_job_completed}
\title{Has a job completed?}
\usage{
has_job_completed(job_id)
}
\arguments{
\item{job_id}{eight-digit job ID given by Peregrine upon
submission.}
}
\value{
logical
}
\description{
Is the job status `COMPLETED`?
}
\author{
Théo Pannetier
}
