% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_comrad_hpc.R
\name{install_comrad_hpc}
\alias{install_comrad_hpc}
\title{Install comrad on the Peregrine HPC}
\usage{
install_comrad_hpc(rebuild = TRUE)
}
\arguments{
\item{rebuild}{logical. If `TRUE`, call `devtools::build()` first.
If `FALSE`, upload and install the source code corresponding to the current
version in `fabrika/libs`.}
}
\description{
Install or update [comrad](https://github.com/TheoPannetier/comrad) from
branch `brute_force` on the Peregrine HPC. The code on this branch is
compiled with flag `-march=native`, and so must be installed from the source
code. This function builds, upload and install the source code to Peregrine.
}
\author{
Théo Pannetier
}
