% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_comsie_hpc.R
\name{install_comsie_hpc}
\alias{install_comsie_hpc}
\title{Install `comsie` on the Peregrine HPC}
\usage{
install_comsie_hpc(rebuild = TRUE)
}
\arguments{
\item{rebuild}{logical. If `TRUE`, call `devtools::build()` first.
If `FALSE`, upload and install the source code corresponding to the current
version in `fabrika/libs`.}
}
\description{
Install `comsie` on the Peregrine HPC
}
\author{
Théo Pannetier
}
