% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_DDD_hpc.R
\name{install_DDD_hpc}
\alias{install_DDD_hpc}
\title{Install `DDD` on the Peregrine HPC}
\usage{
install_DDD_hpc(rebuild = TRUE)
}
\arguments{
\item{rebuild}{logical. If `TRUE`, call `devtools::build()` first.
If `FALSE`, upload and install the source code corresponding to the current
version in `fabrika/libs`.}
}
\description{
Install `DDD` on the Peregrine HPC
}
\author{
Théo Pannetier
}
