% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dd_ml_on_hd.R
\name{is_dd_ml_on_hd}
\alias{is_dd_ml_on_hd}
\alias{is_dd_ml_rds_on_hd}
\alias{is_dd_ml_log_on_hd}
\title{Check if a dd_ml result file is present on my external hard rive}
\usage{
is_dd_ml_rds_on_hd(job_ids, with_fossil = TRUE)

is_dd_ml_log_on_hd(job_ids, with_fossil = TRUE)
}
\arguments{
\item{job_ids}{eight-digit job ID given by Peregrine upon
submission.}
}
\description{
Check if a dd_ml result file is present on my external hard rive
}
\author{
Théo Pannetier
}
