% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dd_ml_on_local.R
\name{is_dd_ml_on_local}
\alias{is_dd_ml_on_local}
\alias{is_dd_ml_rds_on_local}
\alias{is_dd_ml_log_on_local}
\title{Check if a dd_ml result file is present on the local PC}
\usage{
is_dd_ml_rds_on_local(job_ids, with_fossil = TRUE)

is_dd_ml_log_on_local(job_ids)
}
\arguments{
\item{job_ids}{eight-digit job ID given by Peregrine upon
submission.}
}
\description{
Check if a dd_ml result file is present on the local PC
}
\author{
Théo Pannetier
}
