% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sim_on_drive.R
\name{is_sim_on_drive}
\alias{is_sim_on_drive}
\alias{is_sim_csv_on_drive}
\alias{is_sim_log_on_drive}
\title{Check if a simulation file is present on Google Drive}
\usage{
is_sim_csv_on_drive(job_ids)

is_sim_log_on_drive(job_ids)
}
\arguments{
\item{job_ids}{eight-digit job ID given by Peregrine upon
submission.}
}
\description{
Check if a simulation file is present on Google Drive
}
\author{
Théo Pannetier
}
