% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sim_on_hd.R
\name{is_sim_on_hd}
\alias{is_sim_on_hd}
\alias{is_sim_csv_on_hd}
\alias{is_sim_log_on_hd}
\title{Check if a simulation file is present on my external hard rive}
\usage{
is_sim_csv_on_hd(job_ids)

is_sim_log_on_hd(job_ids)
}
\arguments{
\item{job_ids}{eight-digit job ID given by Peregrine upon
submission.}
}
\description{
Check if a simulation file is present on my external hard rive
}
\author{
Théo Pannetier
}
