% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sim_on_hpc.R
\name{is_sim_on_hpc}
\alias{is_sim_on_hpc}
\title{Fetch and print simulation log from Peregrine}
\arguments{
\item{job_id}{eight-digit job ID given by Peregrine upon
submission.}
}
\description{
Given a job ID, prints the corresponding `.log` file from Peregrine.
}
\author{
Théo Pannetier
}
