% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sim_on_local.R
\name{is_sim_on_local}
\alias{is_sim_on_local}
\alias{is_sim_csv_on_local}
\alias{is_sim_log_on_local}
\title{Check if a simulation file is present on the local PC}
\usage{
is_sim_csv_on_local(job_ids)

is_sim_log_on_local(job_ids)
}
\arguments{
\item{job_ids}{eight-digit job ID given by Peregrine upon
submission.}
}
\description{
Check if a simulation file is present on the local PC
}
\author{
Théo Pannetier
}
