% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_id_to_batch_id.R
\name{job_id_to_batch_id}
\alias{job_id_to_batch_id}
\title{Get batch ID from a job ID}
\usage{
job_id_to_batch_id(job_id)
}
\arguments{
\item{job_id}{numeric or character, eight-digit job ID given by Peregrine
upon submission.}
}
\description{
Reads the logbook and returns the batch ID associated with the input job ID.
}
\details{
Try `job_id %>% job_id_to_batch_id() %>% batch_id_to_job_ids()` to
find all other jobs submitted along with `job_id`.
}
\author{
Théo Pannetier
}
