% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_ids_from_string.R
\name{job_ids_from_string}
\alias{job_ids_from_string}
\title{Extract job IDs from a string}
\usage{
job_ids_from_string(string)
}
\arguments{
\item{string}{a string, typically some console output from [queue()] or
[run_comrad_sim_hpc()].}
}
\description{
Returns 8-digits sequences from a string
}
\author{
Théo Pannetier
}
