% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_info.R
\name{job_info}
\alias{job_info}
\title{Call `jobinfo` on Peregrine}
\usage{
job_info(job_id)
}
\arguments{
\item{job_id}{eight-digit job ID given by Peregrine upon
submission.}
}
\description{
Prints the `jobinfo` for a given job ID.
}
\author{
Théo Pannetier
}
