% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_runtime.R
\name{job_runtime}
\alias{job_runtime}
\title{Get job runtime on Peregrine}
\usage{
job_runtime(job_id)
}
\arguments{
\item{job_id}{eight-digit job ID given by Peregrine upon
submission.}
}
\description{
Extract the time a job has been running on Peregrine from `jobinfo`
}
\author{
Théo Pannetier
}
