% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_status.R
\name{job_status}
\alias{job_status}
\title{Get job status on Peregrine}
\usage{
job_status(job_id)
}
\arguments{
\item{job_id}{eight-digit job ID given by Peregrine upon
submission.}
}
\description{
Extract job status from `jobinfo`
}
\author{
Théo Pannetier
}
