% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_to_log.R
\name{path_to_log_hd}
\alias{path_to_log_hd}
\title{Path to a simulation log file on external hard drive}
\usage{
path_to_log_hd(job_id)
}
\arguments{
\item{job_id}{eight-digit job ID given by Peregrine upon submission.}
}
\description{
Returns the absolute path to the .log file corresponding to `job_id` on my
external hard drive.
}
\author{
Theo Pannetier
}
