% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_to_log.R
\name{path_to_log_hpc}
\alias{path_to_log_hpc}
\title{Path to a simulation log file on Peregrine}
\usage{
path_to_log_hpc(job_id, pkg = "comrad")
}
\arguments{
\item{job_id}{eight-digit job ID given by Peregrine upon submission.}
}
\description{
Returns the absolute path to the .log file corresponding to `job_id` on Peregrine
}
\author{
Theo Pannetier
}
