% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_to_log.R
\name{path_to_log_local}
\alias{path_to_log_local}
\title{Path to a simulation log file on local computer}
\usage{
path_to_log_local(job_id)
}
\arguments{
\item{job_id}{eight-digit job ID given by Peregrine upon submission.}
}
\description{
Returns the (local) absolute path to the .log file corresponding to `job_id`
}
\author{
Theo Pannetier
}
