% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_to_sim.R
\name{path_to_sim_hd}
\alias{path_to_sim_hd}
\title{Path to a simulation data file on external hard drive}
\usage{
path_to_sim_hd(job_id)
}
\arguments{
\item{job_id}{eight-digit job ID given by Peregrine upon submission.}
}
\description{
Returns the absolute path to the .csv file corresponding to `job_id` on an
external hard drive
}
\author{
Theo Pannetier
}
