% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_to_sim.R
\name{path_to_sim_local}
\alias{path_to_sim_local}
\title{Path to a simulation data file on local computer}
\usage{
path_to_sim_local(job_id)
}
\arguments{
\item{job_id}{eight-digit job ID given by Peregrine upon submission.}
}
\description{
Returns the (local) absolute path to the .csv file corresponding to `job_id`
}
\author{
Theo Pannetier
}
