% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aicw.R
\name{plot_aicw}
\alias{plot_aicw}
\title{Plot DD models AIC weights as a barplot}
\usage{
plot_aicw(siga, sigk)
}
\arguments{
\item{siga}{parameter `competition_sd` of `comrad`}

\item{sigk}{parameter `carrying_cap_sd` of `comrad`}
}
\description{
Plot DD models AIC weights as a barplot
}
