% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ml_rates.R
\name{plot_ml_rates}
\alias{plot_ml_rates}
\title{Plot maximum likelihood estimates of DD rates}
\usage{
plot_ml_rates(rates_tbl)
}
\arguments{
\item{rates_tbl}{a data frame with cols `N`, `rate`, `dd_model` and `value`.
E.g. the output of `get_ml_rates()`.}
}
\description{
Plot maximum likelihood estimates of DD rates
}
