% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logbook.R
\name{read_logbook}
\alias{read_logbook}
\title{Read logbook}
\usage{
read_logbook(which_one = "sims", pkg = pkg)
}
\arguments{
\item{which_one}{character, which logbook? can be `"sims"` or `"dd_ml_without_fossil"`}
}
\description{
Load `logbook.csv` into R as a data frame
}
\author{
Théo Pannetier
}
