% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ml_res_best.R
\name{read_ml_res_best}
\alias{read_ml_res_best}
\title{Read best maximum likelihood results}
\usage{
read_ml_res_best(siga, sigk)
}
\arguments{
\item{siga}{parameter `competition_sd` of `comrad`}

\item{sigk}{parameter `carrying_cap_sd` of `comrad`}
}
\description{
Returns a table with best maximum likelihood results for each of the 6 DD
models corresponding to parameters `siga` and `sigk`.
}
