% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ml_with_fossil.R
\name{read_ml_with_fossil}
\alias{read_ml_with_fossil}
\title{Read and combine `dd_ML` results}
\usage{
read_ml_with_fossil(job_ids)
}
\arguments{
\item{job_ids}{8-digit codes (chr or num) corresponding to the ID of a
`dd_ML` job, should be one of the entries in
`logbook_dd_ml_with_fossil.csv`.}
}
\description{
Read all `.rds` files corresponding to `job_ids`, combine them and adds
AIC scores.
}
