% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mle_tbl.R
\name{read_mle_tbl}
\alias{read_mle_tbl}
\title{Read all maximum likelihood estimates}
\usage{
read_mle_tbl(siga, sigk)
}
\arguments{
\item{siga}{parameter `competition_sd` of `comrad`}

\item{sigk}{parameter `carrying_cap_sd` of `comrad`}
}
\description{
Returns a table with all maximum likelihood results (for all of the 6 DD
models and all initial parameter sets) corresponding to parameters `siga`
and `sigk`
}
