% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_sampling_hpc.R
\name{reduce_sampling_hpc}
\alias{reduce_sampling_hpc}
\title{Reduce the sampling of a simulated community}
\usage{
reduce_sampling_hpc(job_id, new_sampling_frac = 0.1)
}
\arguments{
\item{job_id}{same thing as usual}

\item{new_sampling_frac}{the new level of sampling}
}
\description{
Reads the `.csv` file corresponding to `job_id`, and samples a fraction of
individuals every generation up to the new sampling fraction. It is assumed
the previous sampling fraction was 0.5
}
