% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime_to_hms.R
\name{runtime_to_hms}
\alias{runtime_to_hms}
\title{Convert Peregrine time format to HMS}
\usage{
runtime_to_hms(runtime)
}
\arguments{
\item{runtime}{a job duration in format `(DD-)HH:MM:SS`}
}
\description{
Convert Peregrine time format to HMS
}
\author{
Theo Pannetier
}
