% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_current_div_hpc.R
\name{summary_last_gen_hpc}
\alias{summary_last_gen_hpc}
\title{Summary of the last generation on the HPC}
\usage{
summary_last_gen_hpc(job_ids)
}
\arguments{
\item{job_id}{eight-digit job ID given by Peregrine upon
submission.}
}
\description{
Read the community data for a (possibly running) comrad simulation on
Peregrine, returns the generation number (t), species diversity (d), and
number of individuals (n) at the last generation.
}
\author{
Theo Pannetier
}
