% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upadate_dd_model_name_rds.R
\name{update_dd_model_name_rds}
\alias{update_dd_model_name_rds}
\title{Replace "x2" model namesto "p" names in .rds files}
\usage{
update_dd_model_name_rds(job_id, with_fossil)
}
\description{
Replace "x2" model namesto "p" names in .rds files
}
