% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_logbook.R
\name{update_logbook}
\alias{update_logbook}
\title{Update logbook entries after job completion}
\usage{
update_logbook(job_ids)
}
\arguments{
\item{job_ids}{eight-digit job IDs given by Peregrine upon
submission.}
}
\description{
`status` and `runtime` entries can be written only after a job has completed
(or failed). This function download the logbook from Peregrine, writes the
entries for `job_ids`, and uploads back to Peregrine.
}
\author{
Théo Pannetier
}
