% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_logbook_drive.R
\name{update_logbook_drive}
\alias{update_logbook_drive}
\title{Update the logbook on Google Drive}
\usage{
update_logbook_drive()
}
\description{
Update the logbook on Google Drive
}
\author{
Théo Pannetier
}
