% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_dd_ml_bash_hpc.R
\name{upload_dd_ml_bash_hpc}
\alias{upload_dd_ml_bash_hpc}
\title{Upload `run_dd_ml.bash` to Peregrine}
\usage{
upload_dd_ml_bash_hpc()
}
\description{
Shortcut to upload/update the script.
}
\author{
Théo Pannetier
}
