% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_rscript_hpc.R
\name{upload_rscript_hpc}
\alias{upload_rscript_hpc}
\title{Upload a fabrika R script to Peregrine}
\usage{
upload_rscript_hpc(path_to_script)
}
\description{
Shortcut to upload/update the script.
}
\author{
Théo Pannetier
}
