% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_sim_drive.R
\name{upload_sim_drive}
\alias{upload_sim_drive}
\alias{upload_sim_csv_drive}
\alias{upload_sim_log_drive}
\title{Upload simulation files to Google Drive}
\usage{
upload_sim_csv_drive(job_ids, pkg = "comrad")

upload_sim_log_drive(job_ids, pkg = "comrad")
}
\arguments{
\item{job_ids}{8-digit numeric or character vector. Peregrine job IDs
identifying the files to download.}
}
\description{
Given a vector of job IDs, will upload the corresponding `.csv` or `.log`
files from the local computer to my Google Drive account.
}
\author{
Théo Pannetier
}
