% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_path_to_fabrika.R
\name{write_path_to_fabrika_local}
\alias{write_path_to_fabrika_local}
\title{Overwrite `path_to_fabrika_local`}
\usage{
write_path_to_fabrika_local()
}
\description{
Calling this function will overwrite the absolute path to the `fabrika`
folder written in [path_to_fabrika_local()], then re-install the package.
Working dir must be set to the fabrika folder before calling this.
}
\author{
Theo Pannetier
}
