% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_logbook_hpc.R
\name{download_logbook_hpc}
\alias{download_logbook_hpc}
\title{Download the logbook from Peregrine to local}
\usage{
download_logbook_hpc()
}
\description{
er, it's all in the title, really.
}
\author{
Théo Pannetier
}
