% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ml_rates.R
\name{get_ml_rates}
\alias{get_ml_rates}
\title{Get maximum likelihood estimates of DD rates}
\usage{
get_ml_rates(aic_tbl)
}
\arguments{
\item{aic_tbl}{a data frame containing ML estimates for the 6 DD models
(*one* row per model) with at least cols `dd_model`, `ml_lambda_0`, `ml_mu_0`,
`ml_k` and `ml_alpha`. E.g. the output of `read_mle_tbl() %>% filter_aic_best()`}
}
\description{
Get maximum likelihood estimates of DD rates
}
