% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ml_rates.R
\name{get_ml_rates}
\alias{get_ml_rates}
\title{Get maximum likelihood estimates of DD rates}
\usage{
get_ml_rates(siga, sigk)
}
\arguments{
\item{siga}{parameter `competition_sd` of `comrad`}

\item{sigk}{parameter `carrying_cap_sd` of `comrad`}
}
\description{
Get maximum likelihood estimates of DD rates
}
