% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_comrad_hpc.R
\name{install_comrad_hpc}
\alias{install_comrad_hpc}
\title{Install comrad on the Peregrine HPC}
\usage{
install_comrad_hpc(ref = "master")
}
\arguments{
\item{ref}{see [remotes::install_github()]}
}
\description{
Install or update the [comrad](https://github.com/TheoPannetier/comrad)
package (from the `master` branch) to Peregrine.
}
\author{
Théo Pannetier
}
