% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_comrad_hpc.R
\name{install_comrad_hpc}
\alias{install_comrad_hpc}
\title{Install comrad on the Peregrine HPC}
\usage{
install_comrad_hpc()
}
\description{
Install or update the package from the `master` branch
([https://github.com/TheoPannetier/comrad]) to Peregrine.
}
\author{
Théo Pannetier
}
