% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aicw.R
\name{plot_aicw}
\alias{plot_aicw}
\title{Plot DD models AIC weights as a barplot}
\usage{
plot_aicw(aic_tbl)
}
\arguments{
\item{aic_tbl}{a data frame containing AIC weights for the 6 DD models
(*one* row per model) with at least cols `dd_model` and `aicw`.
E.g. the output of `read_mle_tbl() %>% filter_aic_best()`.}
}
\description{
Plot DD models AIC weights as a barplot
}
