% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ml_rates.R
\name{plot_ml_rates}
\alias{plot_ml_rates}
\title{Plot maximum likelihood estimates of DD rates}
\usage{
plot_ml_rates(siga, sigk)
}
\arguments{
\item{siga}{parameter `competition_sd` of `comrad`}

\item{sigk}{parameter `carrying_cap_sd` of `comrad`}
}
\description{
Plot maximum likelihood estimates of DD rates
}
