% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_comrad_sim_hpc.R
\name{run_comrad_sim_hpc}
\alias{run_comrad_sim_hpc}
\title{Run comrad simulations on the Peregrine HPC}
\usage{
run_comrad_sim_hpc(
  nb_gens,
  nb_replicates = 1,
  comrad_params = fabrika::create_comrad_params()
)
}
\arguments{
\item{nb_gens}{integer, how many generations should the simulation be run
for?}

\item{nb_replicates}{numeric, number of replicate simulations to run. All
replicates share the same parameters. One job is submitted per replicate.}

\item{comrad_params}{a list of parameters for [comrad::run_simulation()],
as created with [create_comrad_params()]}
}
\description{
Calls `run_comrad_sim.bash` and passes arguments to
[comrad::run_simulation()], submits the simulations through [ssh].
}
\author{
Théo Pannetier
}
